BASIC: colour the plane, colour the tiles, and list what was typed (phase B4.5)
The last of the feature work, so that B5 is only measurement and data.

FILTER now tints. The colour path grew a stage: attribute codes go through the
52 colour table and then through the tint, and the two things that decide how
that looks -- FILTER_TINT and apply_filter -- sit next to each other so a
recording of the real machine replaces them without touching anything else. It
mixes a quarter of the way towards the tint, which is a guess at "lightly
coloured"; the sprite palette skips the stage, the effect being a BG one.

Tiles are 2bpp. Glyphs are read through one accessor now, so callers no longer
care what the source was: the text font stays 1bpp in rodata (letters need no
second colour) and is promoted on read, table A is 2bpp, and a loaded BMP is
whatever the artist drew -- 1bpp ink lands on index 3, which is where single
colour artwork has always been drawn. PALET keeps all three colours of a group
instead of the third alone, PALET S reaches the sprite plane through its own
hook, and a sprite image is repainted when its palette moves because the colours
are baked into it. The placeholder tiles use all three indices (frame, body,
corner mark) so a glance at the screen says whether any of this is working.

LIST and SAVE write back what was typed. A hex literal now crunches to its own
token carrying the same two bytes, so &H6000 lists as &H6000 rather than 24576,
and LOAD? no longer lists as LOAD PRINT (the '?' is PRINT's shorthand, which is
how it got there). Proving that costs one flag: BASIC_ROUNDTRIP makes the runner
list the program, load the listing into a second interpreter and list again --
450_list_roundtrip covers hex, LOAD? and abbreviations.

known_differences.md places every open question from B0 onward into one of four
buckets: settled, an implementation difference we are keeping, waiting on data
from the real machine, or out of scope. Nothing is left unfiled, and each
waiting row names the table to edit. That document is what makes "the features
are done" a statement rather than a feeling.

Golden 103, simulation screens 15, all six sample programs spawn and run clean.
Pool use and stack high-water for two representative programs are in the report
as the baseline B5 compares the device against.

+++ components/basic/assets/basic_tile_a.h
@@ -17,7 +17,7 @@ extern "C" {
 #endif
 
 /// Sprite tile bitmaps, indexed by table A character code.
-extern const uint8_t basic_tile_a[256][8];
+extern const uint16_t basic_tile_a[256][8];
 
 #ifdef __cplusplus
 }
